% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_pkg_data}
\alias{my_pkg_data}
\title{World Health Organization TB data}
\format{
\subsection{\code{my_pkg_data}}{

A data frame with 7,240 rows and 60 columns:
\describe{
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
my_pkg_data
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
